rename + reorganize + react-router install
