+++ package.json
@@ -32,7 +32,8 @@
     "dotenv": "^16.3.1",
     "express": "^4.18.2",
     "react": "^18.2.0",
-    "react-dom": "^18.2.0"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.27.0"
   },
   "devDependencies": {
     "@babel/core": "^7.21.4",
