--- src/App.scss
@@ -1,12 +1,0 @@
-// @import 'variables';
-
-// @font-face {
-//   font-family: 'Ransom';
-//   src: url(assets/Ransom.ttf) format('truetype');
-// }
-
-// * {
-//   margin: 0;
-//   padding: 0;
-//   box-sizing: border-box;
-// }
